# packages/openapi/src/openapi.yaml
openapi: 3.0.3
info:
  title: PeridotID API
  version: 0.1.0
  description: |
    Gaming Identity Platform — Authentication, Identity, and Profile for the Peridot ecosystem.

    ## Base URL
    All endpoints are prefixed with `/v1`.

    | Environment | Base URL |
    |---|---|
    | Production | `https://api.pid.peridotvault.com/v1` |
    | Local dev | `http://localhost:3301/v1` |

    ## Authentication
    Authentication is **cookie-based**. After a successful Google sign-in the API sets two
    `HttpOnly` cookies:

    - `pid_access` — short-lived JWT access token (15 min), sent on every request.
    - `pid_refresh` — long-lived rotating refresh token (30 days), only sent to `/v1/auth`.

    Browsers handle these automatically; non-browser clients must send
    `Cookie: pid_access=<token>` on each request and call `/v1/auth/refresh` before the
    access token expires.
  contact:
    name: PeridotID
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pid.peridotvault.com/v1
    description: Production
  - url: http://localhost:3301/v1
    description: Local development
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Start Google login
      description: |
        Returns the URL a client should navigate to in order to begin Google OAuth.
        The URL points to `GET /auth/google`, which redirects the browser to Google's
        consent screen. After the user consents, the browser is redirected back through
        `/auth/google/callback` and session cookies are set.

        For cross-origin SSO, pass `returnTo` (must be allowlisted, or registered to
        your app via `clientId`) — the callback then redirects to
        `returnTo?pid_code=...` with a one-time exchange code instead of the wallet.

        First-time users always land on the PID picker after authenticating
        (`POST /v1/auth/claim` creates the permanent `<handle>@pid` identity) —
        handles are only ever chosen there, never up front. Returning logins
        continue straight to a session.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                returnTo:
                  type: string
                  format: uri
                  description: Cross-origin URL to return to with a pid_code
                clientId:
                  type: string
                  description: Registered app id (binds the pid_code to your app)
            examples:
              sso:
                summary: Third-party SSO login
                value:
                  returnTo: "https://mygame.dev/callback"
                  clientId: "pidapp_0123456789abcdef0123456789abcdef"
      responses:
        "200":
          description: Google authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                url: "http://localhost:3301/v1/auth/google"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/pid/available:
    get:
      tags: [auth]
      summary: Check PID handle availability
      description: |
        Returns whether `handle` is free (`<handle>@pid`). Used by the onboarding
        handle picker before starting Google sign-up. Handles are lowercase
        `^[a-z0-9_]{3,20}$`; the full PID is `<handle>@pid` — permanent, never
        changeable, reused, or reassigned.
      parameters:
        - name: handle
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9_]{3,20}$"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available:
                    type: boolean
                  pid:
                    type: string
                    nullable: true
                    description: The full PID when available
                    example: "ifal@pid"
              example:
                available: true
                pid: "ifal@pid"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/google:
    get:
      tags: [auth]
      summary: Redirect to Google OAuth consent screen
      description: |
        Redirects (HTTP 302) to Google's OAuth consent screen with `scope=profile email`.

        **Browser-navigation only.** Open the URL in a browser tab (top-level navigation);
        it cannot be called via `fetch()` or the Swagger "Send" button — the cross-origin
        redirect to Google is blocked by CORS and appears as `Failed to fetch`, which is expected.

        Not meant to be called directly by clients — use `POST /auth/login` to obtain the URL.
      responses:
        "302":
          description: Redirect to Google consent screen
        "503":
          $ref: "#/components/responses/GoogleNotConfigured"
  /auth/google/callback:
    get:
      tags: [auth]
      summary: Google OAuth callback
      description: |
        Handles the redirect back from Google. On success it:

        1. Returning credential (`provider, providerUserId` already linked) → issues
           `pid_access` and `pid_refresh` cookies and redirects (with `pid_code`
           when the login carried an allowlisted `returnTo`).
        2. New credential → mints a single-use claim ticket, sets the
           `pid_claim` cookie, and redirects to `CLIENT_SUCCESS_URL?claim=1` for
           the PID picker. Nothing is created until `POST /auth/claim`.

        Both cookies are `HttpOnly; SameSite=Lax`. In production the `Secure` flag is set.

        Strategy/transport failures (stale code, exchange error) redirect to
        `CLIENT_SUCCESS_URL?error=oauth_failed` instead of an error page — the
        wallet shows an inline retry. Cause stays in server logs only.

        **Browser-navigation only.** Google redirects the browser here; do not call it
        directly via `fetch()`.
      responses:
        "302":
          description: Redirect to CLIENT_SUCCESS_URL (or returnTo with pid_code) with session cookies set
        "503":
          $ref: "#/components/responses/GoogleNotConfigured"
  /auth/claim/status:
    get:
      tags: [auth]
      summary: Pending PID claim, if any
      description: |
        Returns the pending post-auth claim from the `pid_claim` cookie
        (verified credential, no identity yet) for the claim picker UI.
      responses:
        "200":
          description: Claim status
          content:
            application/json:
              schema:
                type: object
                required: [pending]
                properties:
                  pending:
                    type: boolean
                  email:
                    type: string
                    nullable: true
                  displayName:
                    type: string
                    nullable: true
                  avatarUrl:
                    type: string
                    nullable: true
              example:
                pending: true
                email: "new@example.com"
                displayName: "New User"
                avatarUrl: null
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/claim:
    post:
      tags: [auth]
      summary: Claim a pending credential under a fresh PID handle
      description: |
        One transaction creates the identity (`<handle>@pid`, permanent) + profile
        + credential and consumes the single-use claim ticket. Issues the session
        cookies. When the claim came from a relying-party login, the response also
        carries the SSO `pidCode` plus the validated `redirectTo` — the hosted page
        navigates there itself. Expired tickets answer `410` — sign in again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [handle]
              properties:
                handle:
                  type: string
                  pattern: "^[a-z0-9_]{3,20}$"
                  description: Permanent handle for `<handle>@pid`
            example:
              handle: "ifal"
      responses:
        "200":
          description: Claimed identity (+ SSO code and return target when applicable)
          content:
            application/json:
              schema:
                type: object
                required: [ok, pid]
                properties:
                  ok:
                    type: boolean
                  pid:
                    type: string
                    pattern: "^[a-z0-9_]{3,20}@pid$"
                    example: "ifal@pid"
                  pidCode:
                    type: string
                    description: One-time SSO code when the claim came from a relying-party login
                  redirectTo:
                    type: string
                    format: uri
                    description: Validated return target for the pidCode (navigate with ?pid_code=)
              example:
                ok: true
                pid: "ifal@pid"
        "400":
          description: Bad handle, or no pending claim (stale page or already claimed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                statusCode: 400
                message: "No pending claim — sign in again to get a fresh one."
        "409":
          description: Handle taken or credential/email already linked
        "410":
          description: Claim ticket expired — sign in again for a fresh one
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/exchange:
    post:
      tags: [auth]
      summary: Exchange a one-time SSO code for the identity
      description: |
        Consumes a `pid_code` issued by a Google or passkey login with `returnTo`
        (optionally bound to a `clientId`) and returns the PeridotID identity.
        Codes are single-use and expire after ~5 minutes. Call server-to-server;
        when the code is bound to an app, the same `clientId` must be presented.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: One-time pid_code from the redirect URL
                clientId:
                  type: string
                  description: Must match the client_id the code was issued for
                clientSecret:
                  type: string
                  description: Required only when the bound app has a secret set (backend-only, never ship in frontend code)
            example:
              code: "mF8xQ2vT9nR4kL7pZ3wX6yA1bD5eF0gH"
              clientId: "pidapp_0123456789abcdef0123456789abcdef"
      responses:
        "200":
          description: The PeridotID identity behind the code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /auth/token:
    post:
      tags: [auth]
      summary: Client-credentials token for an app backend
      description: |
        Machine-to-machine token bound to the app owner's pid, carrying the
        app's clientId. Use `Authorization: Bearer <token>`. `clientSecret` is
        required when the app has one set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId]
              properties:
                clientId:
                  type: string
                clientSecret:
                  type: string
      responses:
        "200":
          description: Access token + TTL
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/authorize:
    post:
      tags: [auth]
      summary: Mint a pid_code for the current session
      description: |
        Issues a one-time pid_code for the caller's existing session — no
        re-authentication. Powers consent screens on PeridotID-hosted pages: the
        user approves, the page redirects itself to returnTo with the code.
        Cookie-authenticated (same-site pages only). returnTo follows the same
        rules as login (loopback, registered app origins, global allowlist).
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [returnTo]
              properties:
                returnTo:
                  type: string
                  format: uri
                clientId:
                  type: string
                  description: Binds the issued code to the app
      responses:
        "200":
          description: The minted code
          content:
            application/json:
              schema:
                type: object
                required: [pidCode]
                properties:
                  pidCode:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token
      description: |
        Rotates the refresh token and issues a fresh access token. The current `pid_refresh`
        cookie is **invalidated** and replaced with a new one (rotation), so each refresh token
        can be used exactly once.

        - 401 if the refresh token is missing, expired, already rotated, or forged.
        - Rate limited to 10 requests / minute.
      responses:
        "204":
          description: New access + refresh cookies set; old refresh token revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/logout:
    post:
      tags: [auth]
      summary: Log out
      description: |
        Revokes the current refresh token and clears both session cookies. Idempotent — safe to
        call when already logged out.
      responses:
        "200":
          description: Session ended and cookies cleared
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pid_access
      description: "JWT access token set by /auth/google/callback or /auth/refresh"
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 400
            message: ["handle must be 3-20 chars: lowercase letters, numbers, underscore"]
...
    Unauthorized:
      description: Missing, expired, or invalid access/refresh token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 401
            message: "Unauthorized"
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 429
            message: "ThrottlerException: Too Many Requests"
    GoogleNotConfigured:
      description: Google OAuth credentials are not set on the server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 503
            message: "Google OAuth is not configured. Set GOOGLE_CLIENT_ID/GOOGLE_CLIENT_SECRET (and GOOGLE_CALLBACK_URL) in the env for this environment"
  schemas:
    LoginResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Navigate here to start Google sign-in
      example:
        url: "http://localhost:3301/v1/auth/google"
    ExchangeResult:
      type: object
      required: [pid, identityId, profile, credentials]
      properties:
        pid:
          type: string
          pattern: "^[a-z0-9_]{3,20}@pid$"
          description: Permanent PID of the signed-in user (treat as opaque)
          example: "ifal@pid"
        identityId:
          type: string
          deprecated: true
          description: Shim — identical to pid, removed next release
          example: "ifal@pid"
        profile:
          type: object
          properties:
            displayName:
              type: string
              nullable: true
            avatarUrl:
              type: string
              nullable: true
        credentials:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              email:
                type: string
                nullable: true
...
    ApiError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string